backend/requirements.txt
django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-celery-beat==2.7.0
djangorestframework-simplejwt==4.8.0
python-dotenv==1.0.1
djongo==1.3.6
pandas==2.2.3
xlrd==2.0.1
setuptools==75.8.2
drf-spectacular==0.28.0
pymongo==3.12.3
dnspython==2.3.0
pytz~=2025.1
PyJWT~=2.10.1
celery~=5.4.0
google-auth
requests
msal
psycopg2-binary
django-celery-beat